% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{drawTogether.allNumeric}
\alias{drawTogether.allNumeric}
\title{Create a panel of 3D scatterplot and corresponding distribution graphs}
\usage{
drawTogether.allNumeric(d, ...)
}
\arguments{
\item{d}{The data frame, containing the numeric values as the first three columns and \code{fill} as RGB codes for colors.}

\item{...}{To be passed to underlying functions.}
}
\description{
Create a panel of 3D scatterplot and corresponding distribution graphs
}
\examples{
drawTogether.allNumeric(testData.allNumeric(100))
}
