% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{drawTogether.allOrdinal}
\alias{drawTogether.allOrdinal}
\title{Create a panel of 3D scatterplot and corresponding distribution graphs}
\usage{
drawTogether.allOrdinal(d, ...)
}
\arguments{
\item{d}{The data frame, containing the factors as the first three columns and \code{fill} as RGB codes for colors.}

\item{...}{To be passed to underlying functions.}
}
\description{
Create a panel of 3D scatterplot and corresponding distribution graphs
}
\examples{
drawTogether.allOrdinal(testData.allOrdinal(100))
}
