% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{getRandomFactors}
\alias{getRandomFactors}
\title{Create a factor vector of random values}
\usage{
getRandomFactors(levels = c("A", "B", "C", "D"), n = 50)
}
\arguments{
\item{levels}{The levels of the factor.}

\item{n}{How many values should the vector have.}
}
\value{
A factor vector.
}
\description{
This is used to test the \code{drawTogether.allOrdinal()} functions and graphs.
}
