% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{onlyScatter.allNumeric}
\alias{onlyScatter.allNumeric}
\title{Create a 3D scatterplot adapted for numeric values}
\usage{
onlyScatter.allNumeric(d, ...)
}
\arguments{
\item{d}{The data frame, containing the numeric values as the first three columns and \code{fill} as RGB codes for colors.}

\item{...}{To be passed to underlying functions.}
}
\description{
Create a 3D scatterplot adapted for numeric values
}
\examples{
onlyScatter.allNumeric(testData.allNumeric(100))
}
