% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{onlyScatter.allOrdinal}
\alias{onlyScatter.allOrdinal}
\title{Create a 3D scatterplot adapted for ordinal values}
\usage{
onlyScatter.allOrdinal(d, ...)
}
\arguments{
\item{d}{The data frame, containing the factors as the first three columns and \code{fill} as RGB codes for colors.}

\item{...}{To be passed to underlying functions.}
}
\description{
Create a 3D scatterplot adapted for ordinal values
}
\examples{
onlyScatter.allOrdinal(testData.allOrdinal(100))
}
