% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{rotateTogether}
\alias{rotateTogether}
\title{Rotate the graph(s) and maybe save them as subsequent png files}
\usage{
rotateTogether(
  saveloc = NA,
  theta = 0,
  phi = 0,
  theta_delta = 0.2,
  phi_delta = 0.1
)
}
\arguments{
\item{saveloc}{If provided, will create a folder in the current working directory and save subsequent frames as separate png files of 1000x1000px.}

\item{theta}{Starting theta angle. See \code{rgl::view3d()}.}

\item{phi}{Starting phi angle. See \code{rgl::view3d()}.}

\item{theta_delta}{Increment with each frame.}

\item{phi_delta}{Increment with each frame. Whenever \code{abs(phi)-89.99>0}, \code{phi_delta<- -phi_delta}.}
}
\description{
This works both with \code{.allOrdinal()} and \code{.allNumeric()} graphs.
}
\examples{
drawTogether.allOrdinal(testData.allOrdinal(100));rotateTogether(theta_delta=2,phi_delta=1)
}
