% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{testData.allOrdinal}
\alias{testData.allOrdinal}
\title{Generate test data for \code{.allOrdinal} functions and graphs}
\usage{
testData.allOrdinal(n = 50)
}
\arguments{
\item{n}{Number of individual data points.}
}
\value{
A \code{data.frame} with 4 columns: \code{x}, \code{y}, \code{z}, and \code{fill}.
}
\description{
Generate test data for \code{.allOrdinal} functions and graphs
}
\examples{
drawTogether.allOrdinal(testData.allOrdinal(100))
}
