% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\docType{data}
\name{vfc.colors}
\alias{vfc.colors}
\title{A set of 6 standard colors to use}
\format{
An object of class \code{character} of length 6.
}
\usage{
vfc.colors
}
\description{
A set of 6 standard colors to use
}
\examples{
\dontrun{ggplot()+scale_fill_manual(values=vfc.colors[1:3])}
}
\keyword{datasets}
